Fix email python_compat verifier contract and frontend API URLs
- Replace nested factory package with flat backend/app.py + models.py
- Export module-level app, db, User, Email, Attachment for verifier imports
- Make email startup smoke check fatal (tests TYPE-A import contract)
- Change all email/todolist frontend API base URLs to relative /api
- Add email/todolist npm-install log paths for consistency
- Ignore frontend build artifacts in .gitignore
