--- mock-platform/python_compat/email-app/backend/app/models/__init__.py
@@ -1,7 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-
-from app.models.user import User
-from app.models.email import Email
-from app.models.attachment import Attachment
