--- mock-platform/python_compat/email-app/backend/app/models/attachment.py
@@ -1,25 +1,0 @@
-from datetime import datetime
-
-from app.models import db
-
-
-class Attachment(db.Model):
-    __tablename__ = "attachments"
-
-    id = db.Column(db.Integer, primary_key=True)
-    email_id = db.Column(db.Integer, db.ForeignKey("emails.id", ondelete="CASCADE"), nullable=True)
-    filename = db.Column(db.String(255), nullable=False)
-    original_filename = db.Column(db.String(255), nullable=False)
-    file_path = db.Column(db.String(500), nullable=False)
-    file_size = db.Column(db.Integer, nullable=False)
-    mime_type = db.Column(db.String(100), nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.utcnow)
-
-    def to_dict(self):
-        return {
-            "id": self.id,
-            "original_filename": self.original_filename,
-            "file_size": self.file_size,
-            "mime_type": self.mime_type,
-            "created_at": self.created_at.isoformat(),
-        }
