--- mock-platform/python_compat/email-app/backend/app/models/user.py
@@ -1,36 +1,0 @@
-from datetime import datetime
-
-from werkzeug.security import check_password_hash, generate_password_hash
-
-from app.models import db
-
-
-class User(db.Model):
-    __tablename__ = "users"
-
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(80), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    password_hash = db.Column(db.String(255), nullable=False)
-    created_at = db.Column(db.DateTime, default=datetime.utcnow)
-
-    sent_emails = db.relationship(
-        "Email", foreign_keys="Email.sender_id", backref="sender", lazy="dynamic"
-    )
-    received_emails = db.relationship(
-        "Email", foreign_keys="Email.recipient_id", backref="recipient", lazy="dynamic"
-    )
-
-    def set_password(self, password):
-        self.password_hash = generate_password_hash(password)
-
-    def check_password(self, password):
-        return check_password_hash(self.password_hash, password)
-
-    def to_dict(self):
-        return {
-            "id": self.id,
-            "username": self.username,
-            "email": self.email,
-            "created_at": self.created_at.isoformat(),
-        }
