+++ tasks/email-writing/environment/email-app/frontend/src/api/api.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-const API_BASE_URL = 'http://localhost:5001/api';
+const API_BASE_URL = '/api';
 
 // Create axios instance with default config
 const api = axios.create({
