+++ tasks/schedule-change-request/environment/todolist-app/frontend/src/services/api.js
@@ -2,7 +2,7 @@
  * API service for communicating with the backend
  */
 
-const API_BASE_URL = 'http://localhost:5002/api';
+const API_BASE_URL = '/api';
 
 /**
  * Generic fetch wrapper with error handling
